Added logic to enable the finish button by default on the Environment tab.
Background:  Previously within the Configure Salesforce wizard, the Finish button was disabled when the user was on the Environment tab unless they picked an identity that had a valid access token.  We required the user to navigate to the next page at which point the finish button would be enabled.  On the page leaving, we prompt the user for their credentials.  This was the wrong behavior.  The finish button should be enabled as long as there aren’t any validation errors within the wizard.  Requiring the user to navigate to the next page was wrong because they weren't actually required to enter anything on the next page.  It is the page leaving which is what we require.  Pressing the finish button would trigger the page leaving as well so there was no reason why it was not enabled.   In this case when the Finish button is pressed, we prompt the user for their credentials and if the user fails to enter valid credentials, an error dialog is displayed and the user is be navigated back to the Environment tab.  This is the same thing that happens when the Next button is pressed.