Fix bug: When scaffolding multiple Salesforce objects, classes are defined multiple times
When scaffolding code from the pre-processed templates, the same IPreprocessedT4Template object is used for each output file of a particular template file (e.g. each object for SalesforceObject.tt).  Because the template's TransformText method appends the text, this causes the each file to contain the content of the previous files.  The fix is to use a new IPreprocessedT4Template each time.