Update Salesforce VSIX ID in call to NuGet API
When I submitted the change to the Salesforce VSIX in commit 74039f53, I forgot to update it in our call to the NuGet API.  I've verified that this is the only other place the VSIX ID is used.

Without this change, the Salesforce handler fails because of an ArgumentNullException when installing the NuGet packages.

In the future, we might consider defining the VSIX ID in such a way that it can be reused between the vsixmanifest and the .NET code.  We might be able to do this in Salesforce.props like for the BuildVersion, but it's not clear how to expose it to the .NET code, so I'm deferring this work.