Use themed checkboxes in the object picker
Now that the bugs in ThemedDialogCheckBoxStyle are fixed, the Salesforce object picker can use it.  The checkbox under this style is larger and thus needs a top margin of 1 and a bottom margin of 2 to have 3 pixels in between adjacent checkboxes, as UX suggested.  UX also suggested making the right margin of the TextBlock 3 pixels, so that when the object has keyboard focus, the focus rectangle doesn't extent as far to the right.

Also, now that ThemedDialogComboBoxStyle defines an ErrorValidation template, the MyDomain ComboBox no longer needs to.