Update Salesforce VSIX ID and name for RTM
For the RTM release, as for the previous public releases, we are updating the VSIX ID because the new VSIX is incompatible with the previous public releases.  This keeps developers using the previous release (RC) from getting a toast notification to update their Salesforce Connected Service, which would fail.

This change also updates the InstallationTarget to Community.  The VSIX team said several months ago that they plan to have a target of "Pro" also support Community for the time being (i.e. Dev14), but that it will change later (at least for Dev15), so it's better to change it now.  The change in InstallationTarget from Pro to Community has already been made in Connected Services Core and the Azure VSIXes.