Ensure all telemetry events are logged by invoking Flush
If the user closes VS shortly after using the Salesforce provider, telemetry events are not logged to App Insights.  More details can be found in https://github.com/developerforce/visual-studio-tools/issues/15

The fix is to invoke TelemetryClient.Flush from the wizard's Dispose method.  ConnectedServiceConfigurator, from which SalesforceConnectedServiceWizard derives, implements IDisposable, so the SalesforceConnectedServiceWizard.Dispose method is invoked when either the wizard is closed or the handler finishes.